Switch to divs, move all key capture to PuzzleView
