+++ src/components/PuzzleCell.tsx
@@ -1,40 +1,14 @@
-import { useEffect, useRef } from "react";
-import { SingleLetter } from "../state/Puzzle";
+import { useRef } from "react";
 import type { PuzzleGameCell } from "../state/State";
 
 interface Props {
   gameCell: PuzzleGameCell;
-  onEnterValue: (value: SingleLetter | "") => void;
   onSelectCell: () => void;
   isSelected: boolean;
 }
 
-const alphaCharacterRegex = /^[A-Za-z]$/;
-
-const getValidInput = (input: string): SingleLetter | "" | null => {
-  if (input.length === 0) {
-    return "";
-  }
-
-  if (input.length === 1 && input[0].match(alphaCharacterRegex)) {
-    return input[0].toUpperCase() as SingleLetter;
-  }
-
-  if (input.length === 2 && input[1].match(alphaCharacterRegex)) {
-    return input[1].toUpperCase() as SingleLetter;
-  }
-
-  return null;
-};
-
 export const PuzzleCell = (props: Props): JSX.Element => {
   const inputRef = useRef<null | HTMLInputElement>(null);
-
-  useEffect(() => {
-    if (inputRef?.current && props.isSelected) {
-      inputRef.current.focus();
-    }
-  }, [props.isSelected, inputRef]);
 
   if (props.gameCell.isBlocked) {
     return <div className="grid-cell" />;
@@ -46,19 +20,14 @@ export const PuzzleCell = (props: Props): JSX.Element => {
 
   return (
     <div className="grid-cell-wrapper">
-      <input
+      <div
         ref={inputRef}
         className={`grid-cell ${inputClassFilled} ${inputClassSelected}`}
         onFocus={props.onSelectCell}
-        onChange={(event) => {
-          const validInput = getValidInput(event.target.value);
-
-          if (validInput !== null) {
-            props.onEnterValue(validInput);
-          }
-        }}
-        value={props.gameCell.filledValue}
-      />
+        onClick={props.onSelectCell}
+      >
+        {props.gameCell.filledValue}
+      </div>
       <p className={`grid-cell-hint-number ${hintClassSelected}`}>{props.gameCell.clueNumber}</p>
     </div>
   );
